refactor: change user profile types to seperate file from user
