+++ src/controllers/user-profile.controller.ts
@@ -1,8 +1,13 @@
 import { db } from '~/db/drizzle';
 import { createAuthRouter } from './router-factory';
-import { getUserAcademic, getUserProfile } from '~/repositories/user-profile.repo';
-import { getUserAcademicRoute, getUserProfileRoute } from '~/routes/user-profile.route';
-import { ErrorSchema, ValidationErrorSchema } from '~/types/responses.type';
+import {
+  getUserAcademic,
+  getUserProfile,
+} from '~/repositories/user-profile.repo';
+import {
+  getUserAcademicRoute,
+  getUserProfileRoute,
+} from '~/routes/user-profile.route';
 
 export const userProfileRoute = createAuthRouter();
 
@@ -18,8 +23,7 @@ userProfileRoute.openapi(getUserAcademicRoute, async (c) => {
 });
 
 userProfileRoute.openapi(getUserProfileRoute, async (c) => {
-  const profile =
-    await getUserProfile(db, c.var.user.id);
+  const profile = await getUserProfile(db, c.var.user.id);
   if (!profile) {
     throw new Error('User not found');
   }
