+++ src/repositories/user-profile.repo.ts
@@ -1,23 +1,21 @@
-import { z } from 'zod';
 import { eq } from 'drizzle-orm';
 import { Database } from '~/db/drizzle';
 import { users } from '~/db/schema';
 import { getCurrentSemesterCodeAndYear } from './course.repo';
 
-
 export async function getUserAcademic(db: Database, userId: string) {
-  var semester
-  const user =  await db.query.users.findFirst({
+  var semester;
+  const user = await db.query.users.findFirst({
     where: eq(users.id, userId),
   });
-  const {semesterCodeTaken, semesterYearTaken} = getCurrentSemesterCodeAndYear()
-  if (semesterCodeTaken === "Genap") {
+  const { semesterCodeTaken, semesterYearTaken } =
+    getCurrentSemesterCodeAndYear();
+  if (semesterCodeTaken === 'Genap') {
     semester = 2 * (semesterYearTaken - user!.angkatan + 1);
   } else {
-      semester = 2 * (semesterYearTaken - user!.angkatan) + 1;
+    semester = 2 * (semesterYearTaken - user!.angkatan) + 1;
   }
-  return {semester, semesterCodeTaken, semesterYearTaken}
-
+  return { semester, semesterCodeTaken, semesterYearTaken };
 }
 
 export async function getUserProfile(db: Database, userId: string) {
